feat: split cli main function
